+++ tests/CommandTest.php
@@ -1,4 +1,21 @@
 <?php
+
+// Fake class
+namespace SebastianBergmann\Git;
+
+class Git
+{
+    protected function execute()
+    {
+        return [getcwd()];
+    }
+
+    public function getDiff($a, $b)
+    {
+        return '';
+    }
+}
+
 namespace tests;
 
 use League\CLImate\CLImate;
@@ -47,49 +64,89 @@ class CommandTest extends \PHPUnit_Framework_TestCase
     public function it_prints_version_info_when_receives_version_argument()
     {
         $arguments = ['php-hound', '--version'];
-
         $cli = $this->getMock(
             'League\CLImate\CLImate',
-            ['out']
+            ['out', 'inline', 'red', 'green', 'yellow']
         );
         $cli->expects($this->once())->method('out')->with('PHP Hound ' . Analyser::VERSION);
-
         $command = new Command($cli, $this->binariesPath, $arguments);
-
         $command->run();
+    }
+
+    /** @test */
+    public function it_limit_results_by_a_git_diff()
+    {
+        // $arguments = ['php-hound', '--git-diff=master..branch'];
+        // $cli = $this->getCliMock();
+        // $command = $this->getMockBuilder(
+        //         'phphound\Command',
+        //         ['getAnalyser']
+        //     )
+        //     ->disableOriginalConstructor()
+        //     ->getMock()
+        // ;
+        // $analyser = $this
+        //     ->getMockBuilder('phphound\Analyser', ['setResultsFilter'])
+        //     ->disableOriginalConstructor()
+        //     ->getMock()
+        // ;
+        // $command
+        //     ->method('getAnalyser')
+        //     ->willReturn($analyser)
+        // ;
+        // $command
+        //     ->expects($this->any())
+        //     ->method('hasArgumentValue')
+        //     ->with('git-diff')
+        //     ->willReturn(true)
+        // ;
+        // $analyser
+        //     ->expects($this->once())
+        //     ->method('setResultsFilter')
+        // ;
+        // $command->run();
     }
 
     /** @test */
     public function it_runs_analysis_tools()
     {
-        $cli = $this->getMock('League\CLImate\CLImate', ['output']);
+        $cli = $this->getCliMock();
         $arguments = ['php-hound', 'src'];
         $command = new Command($cli, $this->binariesPath, $arguments);
         $command->run();
     }
 
     /** @test */
-    public function it_accepts_target_path()
+    public function it_relative_target_path()
     {
-        $cli = $this->getMock('League\CLImate\CLImate', ['output']);
+        $cli = $this->getCliMock();
         $arguments = ['php-hound', 'target/path'];
         $command = new Command($cli, $this->binariesPath, $arguments);
-        $this->assertEquals('target/path', $command->getAnalysedPath());
+        $this->assertEquals([getcwd() . DIRECTORY_SEPARATOR . 'target/path'], $command->getAnalysedPaths());
+    }
+
+    /** @test */
+    public function it_full_target_path()
+    {
+        $cli = $this->getCliMock();
+        $arguments = ['php-hound', DIRECTORY_SEPARATOR . 'target/path'];
+        $command = new Command($cli, $this->binariesPath, $arguments);
+        $this->assertEquals([DIRECTORY_SEPARATOR . 'target/path'], $command->getAnalysedPaths());
     }
 
     /** @test */
     public function it_uses_current_dir_as_target_path_when_not_informed()
     {
-        $cli = $this->getMock('League\CLImate\CLImate', ['output']);
+        $cli = $this->getCliMock();
         $arguments = ['php-hound'];
         $command = new Command($cli, $this->binariesPath, $arguments);
-        $this->assertEquals('.', $command->getAnalysedPath());
+        $this->assertEquals([getcwd() . DIRECTORY_SEPARATOR . '.'], $command->getAnalysedPaths());
     }
 
     /** @test */
     public function it_accepts_ignore_param()
     {
-        $cli = $this->getMock('League\CLImate\CLImate', ['output']);
+        $cli = $this->getCliMock();
         $arguments = ['php-hound', '--ignore=dir'];
 
         $command = new Command($cli, $this->binariesPath, $arguments);
@@ -99,7 +156,7 @@ class CommandTest extends \PHPUnit_Framework_TestCase
     /** @test */
     public function it_accepts_ignore_param_with_multiple_directories()
     {
-        $cli = $this->getMock('League\CLImate\CLImate', ['output']);
+        $cli = $this->getCliMock();
         $arguments = ['php-hound', '--ignore=dir1,dir2'];
 
         $command = new Command($cli, $this->binariesPath, $arguments);
@@ -109,7 +166,7 @@ class CommandTest extends \PHPUnit_Framework_TestCase
     /** @test */
     public function it_has_ignore_param_default()
     {
-        $cli = $this->getMock('League\CLImate\CLImate', ['output']);
+        $cli = $this->getCliMock();
         $arguments = ['php-hound'];
 
         $command = new Command($cli, $this->binariesPath, $arguments);
@@ -119,7 +176,7 @@ class CommandTest extends \PHPUnit_Framework_TestCase
     /** @test */
     public function it_accepts_ignore_param_with_empty_value()
     {
-        $cli = $this->getMock('League\CLImate\CLImate', ['output']);
+        $cli = $this->getCliMock();
         $arguments = ['php-hound', '--ignore='];
 
         $command = new Command($cli, $this->binariesPath, $arguments);
@@ -131,7 +188,7 @@ class CommandTest extends \PHPUnit_Framework_TestCase
     {
         $this->setExpectedException('UnexpectedValueException');
         $arguments = ['php-hound', '--format=invalid'];
-        $cli = new CLImate;
+        $cli = $this->getCliMock();
         $command = new Command($cli, $this->binariesPath, $arguments);
     }
 
@@ -139,7 +196,7 @@ class CommandTest extends \PHPUnit_Framework_TestCase
     public function it_uses_json_output_with_format_json_param()
     {
         $arguments = ['php-hound', '--format=json'];
-        $cli = new CLImate;
+        $cli = $this->getCliMock();
         $command = new Command($cli, $this->binariesPath, $arguments);
 
         $this->assertInstanceOf('phphound\output\JsonOutput', $this->invokeGetOutput($command));
@@ -149,7 +206,7 @@ class CommandTest extends \PHPUnit_Framework_TestCase
     public function it_uses_json_output_with_f_json_param()
     {
         $arguments = ['php-hound', '-f=json'];
-        $cli = new CLImate;
+        $cli = $this->getCliMock();
         $command = new Command($cli, $this->binariesPath, $arguments);
 
         $this->assertInstanceOf('phphound\output\JsonOutput', $this->invokeGetOutput($command));
@@ -159,7 +216,7 @@ class CommandTest extends \PHPUnit_Framework_TestCase
     public function it_uses_xml_output_with_format_xml_param()
     {
         $arguments = ['php-hound', '--format=xml'];
-        $cli = new CLImate;
+        $cli = $this->getCliMock();
         $command = new Command($cli, $this->binariesPath, $arguments);
 
         $this->assertInstanceOf('phphound\output\XmlOutput', $this->invokeGetOutput($command));
@@ -169,7 +226,7 @@ class CommandTest extends \PHPUnit_Framework_TestCase
     public function it_uses_xml_output_with_f_xml_param()
     {
         $arguments = ['php-hound', '-f=xml'];
-        $cli = new CLImate;
+        $cli = $this->getCliMock();
         $command = new Command($cli, $this->binariesPath, $arguments);
 
         $this->assertInstanceOf('phphound\output\XmlOutput', $this->invokeGetOutput($command));
@@ -179,7 +236,7 @@ class CommandTest extends \PHPUnit_Framework_TestCase
     public function it_uses_csv_output_with_format_csv_param()
     {
         $arguments = ['php-hound', '--format=csv'];
-        $cli = new CLImate;
+        $cli = $this->getCliMock();
         $command = new Command($cli, $this->binariesPath, $arguments);
 
         $this->assertInstanceOf('phphound\output\CsvOutput', $this->invokeGetOutput($command));
@@ -189,7 +246,7 @@ class CommandTest extends \PHPUnit_Framework_TestCase
     public function it_uses_csv_output_with_f_csv_param()
     {
         $arguments = ['php-hound', '-f=csv'];
-        $cli = new CLImate;
+        $cli = $this->getCliMock();
         $command = new Command($cli, $this->binariesPath, $arguments);
 
         $this->assertInstanceOf('phphound\output\CsvOutput', $this->invokeGetOutput($command));
@@ -199,7 +256,7 @@ class CommandTest extends \PHPUnit_Framework_TestCase
     public function it_uses_html_output_with_format_html_param()
     {
         $arguments = ['php-hound', '--format=html'];
-        $cli = new CLImate;
+        $cli = $this->getCliMock();
         $command = new Command($cli, $this->binariesPath, $arguments);
 
         $this->assertInstanceOf('phphound\output\HtmlOutput', $this->invokeGetOutput($command));
@@ -209,7 +266,7 @@ class CommandTest extends \PHPUnit_Framework_TestCase
     public function it_uses_html_output_with_f_html_param()
     {
         $arguments = ['php-hound', '-f=html'];
-        $cli = new CLImate;
+        $cli = $this->getCliMock();
         $command = new Command($cli, $this->binariesPath, $arguments);
 
         $this->assertInstanceOf('phphound\output\HtmlOutput', $this->invokeGetOutput($command));
@@ -221,5 +278,13 @@ class CommandTest extends \PHPUnit_Framework_TestCase
         $method = $class->getMethod('getOutput');
         $method->setAccessible(true);
         return $method->invoke($command);
+    }
+
+    protected function getCliMock()
+    {
+        return $this->getMock(
+            'League\CLImate\CLImate',
+            ['out', 'inline', 'red', 'green', 'yellow', 'cyan', 'br']
+        );
     }
 }